Adding Docstrings
