+++ deeplake/integrations/mmdet/converters/pascal_format.py
@@ -1,32 +1,31 @@
-import deeplake.integrations.mmdet.converters.coco_format_converters as coco_format_converters
-from typing import Any, Callable, Dict, Tuple, List
+import numpy as np
 
-BBOX_FORMAT_TO_COCO_CONVERTER: Dict[Tuple[str, str], Callable] = {
-    ("LTWH", "pixel"): lambda x, y: x,
-    ("LTWH", "fractional"): coco_format_converters.coco_frac_2_coco_pixel,
-    ("LTRB", "pixel"): coco_format_converters.pascal_pixel_2_coco_pixel,
-    ("LTRB", "fractional"): coco_format_converters.pascal_frac_2_coco_pixel,
-    ("CCWH", "pixel"): coco_format_converters.yolo_pixel_2_coco_pixel,
-    ("CCWH", "fractional"): coco_format_converters.yolo_frac_2_coco_pixel,
+import deeplake.integrations.mmdet.converters.bbox_format as bbox_format
+import deeplake.integrations.mmdet.converters.pascal_format_converters as pascal_format_converters
+
+
+BBOX_FORMAT_TO_PASCAL_CONVERTER = {
+    ("LTWH", "pixel"): pascal_format_converters.coco_pixel_2_pascal_pixel,
+    ("LTWH", "fractional"): pascal_format_converters.coco_frac_2_pascal_pixel,
+    ("LTRB", "pixel"): lambda x, y: x,
+    ("LTRB", "fractional"): pascal_format_converters.pascal_frac_2_pascal_pixel,
+    ("CCWH", "pixel"): pascal_format_converters.yolo_pixel_2_pascal_pixel,
+    ("CCWH", "fractional"): pascal_format_converters.yolo_frac_2_pascal_pixel,
 }
 
 
-def convert(
-    bbox: List[float], bbox_format: Tuple[str, str], shape: Tuple[int, int]
-) -> List[float]:
+def convert(bbox: np.ndarray, bbox_info: dict, shape: tuple) -> np.ndarray:
     """
-    Convert bounding boxes to COCO format (top-left x, top-left y, width, height)
-    in pixel coordinates using the appropriate converter function based on the provided bbox_format.
+    Converts bounding box coordinates to Pascal format.
 
     Args:
-        bbox (List[float]): Bounding boxes to be converted.
-        bbox_format (Tuple[str, str]): The format of the input bounding boxes.
-            First element of the tuple represents the mode (e.g. 'LTWH', 'LTRB', 'CCWH').
-            Second element of the tuple represents the type (e.g. 'pixel', 'fractional').
-        images (Any): Images associated with the bounding boxes.
+        bbox (np.ndarray): A Numpy array containing bounding box coordinates.
+        bbox_info (dict): A dictionary containing information about the bounding box format. Ex: {"mode": "LTWH", "type": "pixel"}
+        shape (tuple): A tuple containing the shape of the image.
 
     Returns:
-        List[float]: Bounding boxes in COCO format.
+        np.ndarray: A Numpy array containing bounding box coordinates in Pascal format.
     """
-    converter = BBOX_FORMAT_TO_COCO_CONVERTER[bbox_format]
+    bbox_info = bbox_format.get_bbox_format(bbox, bbox_info)
+    converter = BBOX_FORMAT_TO_PASCAL_CONVERTER[bbox_info]
     return converter(bbox, shape)
